+++ transcriptic/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import print_function
-from transcriptic.objects import Run, Project, Aliquot, Resource
+from transcriptic.objects import Run, Project
 from transcriptic.objects import Container, Dataset, ProtocolPreview
 from transcriptic.config import Connection
 from .version import __version__